requirements.txt
google-auth 
google-auth-oauthlib 
google-auth-httplib2 
google-api-python-client
Flask
firebase_admin